Move PipSession helper functions to network.session
This new module is a catch-all for PipSession helpers.
